@@ -20,102 +20,100 @@
     "make:species": "cd src/common/data && node make.js"
   },
   "dependencies": {
-    "@capacitor-community/camera-preview": "^5.0.0",
-    "@capacitor/android": "^5.3.0",
-    "@capacitor/app": "^5.0.6",
-    "@capacitor/camera": "^5.0.7",
-    "@capacitor/core": "^5.3.0",
-    "@capacitor/device": "^5.0.6",
-    "@capacitor/filesystem": "^5.1.3",
-    "@capacitor/geolocation": "^5.0.6",
-    "@capacitor/haptics": "^5.0.6",
-    "@capacitor/ios": "^5.3.0",
-    "@capacitor/keyboard": "^5.0.6",
-    "@capacitor/network": "^5.0.6",
-    "@capacitor/splash-screen": "^5.0.6",
-    "@capacitor/status-bar": "^5.0.6",
-    "@changey/react-leaflet-markercluster": "^4.0.0-rc1",
-    "@flumens/bigu": "^0.4.0",
-    "@flumens/ionic": "1.26.1",
+    "@capacitor-community/camera-preview": "5.0.0",
+    "@capacitor/android": "5.3.0",
+    "@capacitor/app": "5.0.6",
+    "@capacitor/camera": "5.0.7",
+    "@capacitor/core": "5.3.0",
+    "@capacitor/device": "5.0.6",
+    "@capacitor/filesystem": "5.1.3",
+    "@capacitor/geolocation": "5.0.6",
+    "@capacitor/haptics": "5.0.6",
+    "@capacitor/ios": "5.3.0",
+    "@capacitor/keyboard": "5.0.6",
+    "@capacitor/network": "5.0.6",
+    "@capacitor/splash-screen": "5.0.6",
+    "@capacitor/status-bar": "5.0.6",
+    "@flumens/bigu": "0.4.0",
+    "@flumens/ionic": "2.0.0-alpha.24",
     "@ionic-native/core": "5.36.0",
-    "@ionic-native/launch-review": "^5.29.0",
-    "@ionic/core": "^6.2.6",
-    "@ionic/react": "^6.2.6",
-    "@ionic/react-router": "^6.2.6",
-    "@sentry/browser": "^7.16.0",
-    "@sentry/integrations": "^7.16.0",
-    "axios": "^0.27.2",
-    "clsx": "^1.1.1",
-    "cordova-launch-review": "^4.0.1",
-    "cordova-plugin-device": "^2.1.0",
-    "cordova-sqlite-storage": "6.0.0",
-    "deep-diff": "^1.0.2",
-    "formik": "2.2.9",
-    "i18next": "^21.8.9",
-    "ionicons": "^6.0.2",
-    "json-stable-stringify": "^1.0.1",
-    "jwt-decode": "^3",
-    "leaflet": "1.7.1",
-    "leaflet.markercluster": "^1.5.3",
+    "@ionic-native/launch-review": "5.29.0",
+    "@ionic/core": "7.3.3",
+    "@ionic/react": "7.3.3",
+    "@ionic/react-router": "7.3.3",
+    "@sentry/browser": "7.68.0",
+    "@sentry/integrations": "7.68.0",
+    "axios": "1.5.0",
+    "clsx": "2.0.0",
+    "cordova-launch-review": "4.0.1",
+    "cordova-plugin-device": "2.1.0",
+    "cordova-sqlite-storage": "6.1.0",
+    "deep-diff": "1.0.2",
+    "formik": "2.4.4",
+    "i18next": "23.5.1",
+    "ionicons": "7.1.2",
+    "json-stable-stringify": "1.0.2",
+    "jwt-decode": "3",
+    "leaflet": "1.9.4",
     "localforage-cordovasqlitedriver": "1.8.0",
-    "lodash.mergewith": "^4.6.2",
-    "mapbox-gl": "2.4.1",
-    "mobx": "^6.6.0",
-    "mobx-react": "^7.5.0",
-    "mobx-utils": ">=6",
+    "lodash.mergewith": "4.6.2",
+    "mobx": "6.10.2",
+    "mobx-react": "7.5.0",
+    "mobx-utils": "6.0.8",
     "mock-geolocation": "1.0.11",
-    "proj4": "^2",
-    "proj4leaflet": "^1",
-    "prop-types": "15.8.1",
-    "prop-types-exact": "1.2.0",
-    "react": "^17.0.2",
-    "react-autosuggest": "^10",
-    "react-dom": "^17.0.2",
-    "react-easy-crop": "^4.6.2",
-    "react-i18next": "^11.17.1",
-    "react-leaflet": "3.2.5",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-window": "^1.8.7",
-    "swiper": "^8.2.2",
-    "yup": "0.32.11"
+    "proj4": "^2.9.0",
+    "proj4leaflet": "^1.0.2",
+    "react": "18.2.0",
+    "react-autosuggest": "10.1.0",
+    "react-dom": "18.2.0",
+    "react-easy-crop": "5.0.2",
+    "react-i18next": "13.2.2",
+    "react-map-gl": "7.1.6",
+    "react-router": "5.3.4",
+    "react-router-dom": "5.3.4",
+    "react-window": "1.8.9",
+    "swiper": "10.2.0",
+    "yup": "1.3.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.10",
+    "@babel/core": "^7.22.17",
+    "@babel/plugin-proposal-private-property-in-object": "7.21.11",
     "@babel/plugin-syntax-dynamic-import": "7.8.3",
-    "@babel/plugin-transform-exponentiation-operator": "^7.16.7",
-    "@babel/plugin-transform-runtime": "^7.17.10",
-    "@babel/preset-env": "^7.17.10",
-    "@babel/preset-react": "^7.16.7",
-    "@babel/preset-typescript": "^7.16.7",
+    "@babel/plugin-transform-exponentiation-operator": "^7.22.5",
+    "@babel/plugin-transform-runtime": "^7.22.15",
+    "@babel/preset-env": "^7.22.15",
+    "@babel/preset-react": "^7.22.15",
+    "@babel/preset-typescript": "^7.22.15",
     "@capacitor/cli": "^5.3.0",
     "@flumens/eslint-config": "^2.0.1",
-    "@flumens/webpack-config": "^5.0.0",
-    "@types/jest": "^27.5.1",
-    "@types/leaflet": "^1.7.10",
+    "@flumens/prettier-config": "0.2.3",
+    "@flumens/webpack-config": "^5.2.0",
+    "@types/jest": "^29.5.4",
+    "@types/leaflet": "^1.9.4",
     "@types/lodash.mergewith": "^4.6.7",
-    "@types/react": "17.0.24",
-    "@types/react-dom": "17.0.9",
-    "@types/react-router": "^5.1.18",
-    "@types/react-router-dom": "^5.3.3",
+    "@types/proj4": "^2.5.2",
+    "@types/react": "^18.2.28",
+    "@types/react-dom": "^18.2.7",
+    "@types/react-router": "^5.1.20",
+    "@types/react-router-dom": "5.3.3",
     "@types/react-window": "^1.8.5",
     "babel-eslint": "10.1.0",
-    "babel-jest": "^28.1.0",
-    "babel-plugin-module-resolver": "4.1.0",
+    "babel-jest": "^29.7.0",
+    "babel-plugin-module-resolver": "^5.0.0",
     "cordova-res": "^0.15.4",
-    "dotenv": "^16.0.1",
-    "eslint": "^8.15.0",
-    "husky": "^8.0.1",
-    "jest": "^28.1.0",
-    "jest-environment-jsdom": "^28.1.0",
+    "dotenv": "^16.3.1",
+    "eslint": "^8.49.0",
+    "husky": "^8.0.3",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "jest-transform-stub": "2.0.0",
     "sinon": "^8.0.4",
-    "tslib": "^2.4.0",
-    "typescript": "^4.6.4",
-    "vercel": "^24.2.3",
-    "webpack": "^5.72.1",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.9.0"
+    "tslib": "^2.6.2",
+    "typescript": "^5.2.2",
+    "vercel": "^32.2.1",
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   },
   "engines": {
     "node": ">=12",
