@@ -1,21 +1,21 @@
-import ReactDOM from 'react-dom';
-import { setupIonicReact, isPlatform } from '@ionic/react';
-import appModel from 'models/app';
-import userModel from 'models/user';
-import savedSamples from 'models/savedSamples';
-import { StatusBar, Style as StatusBarStyle } from '@capacitor/status-bar';
-import { SplashScreen } from '@capacitor/splash-screen';
-import { App as AppPlugin } from '@capacitor/app';
-import i18n from 'i18next';
-import { initReactI18next } from 'react-i18next';
-import config from 'common/config';
 import { configure as mobxConfig } from 'mobx';
-import { initAnalytics } from '@flumens';
-import App from './App';
+import i18n from 'i18next';
+import { createRoot } from 'react-dom/client';
+import { initReactI18next } from 'react-i18next';
+import { App as AppPlugin } from '@capacitor/app';
 import '@capacitor/core';
+import { SplashScreen } from '@capacitor/splash-screen';
+import { StatusBar, Style as StatusBarStyle } from '@capacitor/status-bar';
+import { initAnalytics } from '@flumens';
 import '@ionic/core/css/core.css';
 import '@ionic/core/css/ionic.bundle.css';
+import { setupIonicReact, isPlatform } from '@ionic/react';
+import config from 'common/config';
 import 'common/theme.scss';
+import appModel from 'models/app';
+import savedSamples from 'models/savedSamples';
+import userModel from 'models/user';
+import App from './App';
 
 console.log('🚩 App starting.'); // eslint-disable-line
 
@@ -45,7 +45,9 @@ async function init() {
   appModel.attrs.appSession += 1;
   appModel.save();
 
-  ReactDOM.render(<App />, document.getElementById('root'));
+  const container = document.getElementById('root');
+  const root = createRoot(container!);
+  root.render(<App />);
 
   if (isPlatform('hybrid')) {
     StatusBar.setStyle({
