@@ -1,8 +1,11 @@
-import { FC } from 'react';
 import { IonBackdrop } from '@ionic/react';
 import './styles.scss';
 
-const CustomAlert: FC = ({ children }) => (
+type Props = {
+  children: any;
+};
+
+const CustomAlert = ({ children }: Props) => (
   <div className="custom-alert">
     <IonBackdrop tappable visible stopPropagation />
     <div className="message">{children}</div>
