@@ -1,169 +1,231 @@
-import { FC, useState, useRef } from 'react';
-import {
-  IonIcon,
-  IonButton,
-  IonModal,
-  IonContent,
-  useIonViewDidEnter,
-  useIonViewWillLeave,
-  IonHeader,
-  IonToolbar,
-  isPlatform,
-} from '@ionic/react';
-import {
-  ModelLocation as ModelLocationOrig,
-  useToast,
-  HandledError,
-} from '@flumens';
-import Sample from 'models/sample';
-import appModel from 'models/app';
+/* eslint-disable no-restricted-syntax */
+import { useState } from 'react';
+import { observer } from 'mobx-react';
 import { Haptics, ImpactStyle } from '@capacitor/haptics';
-import { starOutline } from 'ionicons/icons';
-import L from 'leaflet';
+import {
+  MapHeader,
+  MapSettingsPanel,
+  Page,
+  Main,
+  RadioInput,
+  textToLocation,
+  mapEventToLocation,
+  toggleGPS,
+  isValidLocation,
+  useToast,
+  Location,
+  RadioInputOption,
+} from '@flumens';
+import { isPlatform } from '@ionic/core';
+import { useIonViewWillLeave } from '@ionic/react';
 import config from 'common/config';
-import { GPS_DISABLED_ERROR_MESSAGE } from 'helpers/GPS';
 import locationNameIcon from 'common/images/location-name.svg';
-import PastLocationsList from 'Components/PastLocationsList';
+import appModel from 'models/app';
+import Sample, { getEmptyLocation } from 'models/sample';
+import { hasGPSPermissions } from 'helpers/GPS';
+import LeafletMap from './LeafletMap';
+import osMapIcon from './LeafletMap/osMapIcon.png';
+import MapboxMap from './MapboxMap';
+import satelliteMapIcon from './MapboxMap/satelliteMapIcon.png';
+import standardMapIcon from './MapboxMap/standardMapIcon.png';
+import PastLocationsPanel from './PastLocations';
 import './styles.scss';
 
-type Props = {
-  sample: Sample;
-  subSample?: Sample;
-  locationNameProps?: any;
+export const setModelLocation = async (
+  model: Sample,
+  newLocation: Location
+) => {
+  if (model.isGPSRunning()) model.stopGPS(); // we don't need the GPS running and overwriting the selected location
+
+  const isFromMap = newLocation?.source === 'map';
+  isPlatform('hybrid') &&
+    isFromMap &&
+    Haptics.impact({ style: ImpactStyle.Light });
+
+  Object.assign(
+    model.attrs.location, // carry over name, geocoded etc
+    getEmptyLocation(), // overwrite core location values
+    newLocation
+  );
+
+  model.save();
+
+  if (!isValidLocation(newLocation)) return;
+  appModel.setLocation(model.attrs.location);
 };
 
-const SNAP_POSITIONS = [0, 0.3, 0.5, 1];
-const DEFAULT_SNAP_POSITION = 0.3;
+type Styles = 'satellite' | 'os' | 'os_explorer';
+export const useMapStyles = (): [Styles, any, RadioInputOption[]] => {
+  const layers: RadioInputOption[] = [
+    {
+      value: 'Map Type',
+      isPlaceholder: true,
+    },
+    {
+      label: 'Satellite',
+      value: 'satellite',
+      icon: satelliteMapIcon,
+    },
+    {
+      label: 'Ordnance Survey',
+      value: 'os',
+      icon: standardMapIcon,
+    },
+    {
+      label: 'Ordnance Survey (Explorer)',
+      value: 'os_explorer',
+      icon: osMapIcon,
+    },
+  ];
 
-const ModelLocation: FC<Props> = ({ sample, subSample, ...otherProps }) => {
-  const toast = useToast();
-  const [showPastLocations, setShowPastLocations] = useState(false);
+  const [currentLayer, setCurrentLayer] = useState<Styles>('satellite');
+
+  return [currentLayer, setCurrentLayer, layers];
+};
+
+type Props = {
+  subSample?: any;
+  sample: any;
+  setLocation?: any;
+  skipLocationName?: boolean;
+  skipPastLocations?: boolean;
+};
+
+const ModelLocationMap = ({
+  subSample,
+  sample,
+  setLocation = setModelLocation,
+  skipLocationName,
+  skipPastLocations,
+}: Props) => {
   const model = subSample || sample;
+  const isDisabled = model.isDisabled();
+  const location = model.attrs.location || {};
+  const parentLocation = model.parent?.attrs.location;
 
-  const buttonRef = useRef<any>(null);
-  const fixFavBtnClickMapPropagation = () =>
-    buttonRef.current && L.DomEvent.disableClickPropagation(buttonRef.current);
-  useIonViewDidEnter(fixFavBtnClickMapPropagation);
-
-  const fixOpenModalWhenLeavingPage = () => setShowPastLocations(false);
-  useIonViewWillLeave(fixOpenModalWhenLeavingPage);
-
-  const setLocation = async (loc: any, reset = false) => {
-    let location = loc;
-
-    if (model.isGPSRunning()) model.stopGPS(); // we don't need the GPS running and overwriting the selected location
-
-    if (!reset) {
-      // extend old location to preserve its previous attributes like name or id
-      const oldLocation = model.attrs.location || {};
-      location = { ...oldLocation, ...location };
+  const onManuallyTypedLocationChange = (e: any) => {
+    const value = e?.target?.value;
+    if (!value) {
+      setLocation(model, {});
+      return;
     }
 
-    isPlatform('hybrid') && Haptics.impact({ style: ImpactStyle.Light });
+    const newLocation = textToLocation(value);
+    if (!isValidLocation(newLocation)) return;
 
-    // eslint-disable-next-line no-param-reassign
-    model.attrs.location = location;
-
-    if (location.latitude) appModel.setLocation(location);
-
-    const onError = (error: any) => console.log(error);
-    return model.save().catch(onError);
+    setLocation(model, newLocation);
   };
 
-  const onSelectPastLoaction = (location: any) => {
-    if (sample.isGPSRunning()) sample.stopGPS();
-    setShowPastLocations(false);
-
-    isPlatform('hybrid') && Haptics.impact({ style: ImpactStyle.Light });
-
-    // eslint-disable-next-line no-param-reassign
-    model.attrs.location = location;
-    model.save();
+  const onLocationNameChange = ({ name, geocoded }: any) => {
+    model.attrs.location = { ...model.attrs.location, name, geocoded };
   };
 
-  async function onGPSClick() {
-    try {
-      await ModelLocationOrig.utils.onGPSClick(model);
-    } catch (error: any) {
-      if (error.message === GPS_DISABLED_ERROR_MESSAGE) {
-        toast.error(new HandledError(GPS_DISABLED_ERROR_MESSAGE));
-        return;
-      }
+  const [showSettings, setShowSettings] = useState(false);
+  const onCloseSettings = () => setShowSettings(false);
+  const onLayersClick = () => setShowSettings(!showSettings);
 
-      toast.error(error);
+  const [currentStyle, setCurrentStyle, styles] = useMapStyles();
+  const onStyleChange = (newLayer: string) => {
+    setCurrentStyle(newLayer);
+    setShowSettings(false);
+  };
+
+  const onMapClick = (e: any) => setLocation(model, mapEventToLocation(e));
+
+  const toast = useToast();
+  const onGPSClick = async () => {
+    const hasPermissions = await hasGPSPermissions();
+    if (!hasPermissions) {
+      toast.warn('Location services are not enabled');
+      return;
     }
-  }
+
+    toggleGPS(model);
+  };
+
+  const [showPastLocations, setShowPastLocations] = useState(false);
+  const onPastLocationsClick = () => setShowPastLocations(!showPastLocations);
+
+  const fixOpenPastLocationsOnPageLeave = () => setShowPastLocations(false);
+  useIonViewWillLeave(fixOpenPastLocationsOnPageLeave);
+
+  const fixOpenSettingsOnPageLeave = () => setShowSettings(false);
+  useIonViewWillLeave(fixOpenSettingsOnPageLeave);
+
+  const isMapboxMap = currentStyle !== 'os_explorer';
 
   return (
-    <>
-      <ModelLocationOrig
-        model={model} // eslint-disable-line
-        mapProviderOptions={config.map}
-        useGridRef
-        useGridMap
-        onGPSClick={onGPSClick}
-        backButtonProps={{ text: 'Back' }}
-        setLocation={setLocation}
-        locationNameProps={{
-          onChange: ModelLocationOrig.utils.onLocationNameChange,
-          suggestLocations: appModel.attrs.locations || [],
-          placeholder: 'Site name eg nearby village',
-          icon: locationNameIcon,
-          geocodingParams: {
-            access_token: config.map.mapboxApiKey,
-            types: 'locality,place,district,neighborhood,region,postcode',
-            country: 'GB',
-          },
-        }}
-        className="with-past-locations"
-        {...otherProps}
-      >
-        <div className="leaflet-control-button leaflet-control past-locations-control">
-          <div className="leaflet-buttons-control-button">
-            <IonButton
-              ref={buttonRef}
-              onClick={(e: any) => {
-                L.DomEvent.preventDefault(e);
-                L.DomEvent.stopPropagation(e);
-                e.preventDefault();
-                e.stopPropagation();
+    <Page id="model-location">
+      <MapHeader>
+        <MapHeader.Location
+          location={location}
+          onChange={onManuallyTypedLocationChange}
+          backButtonProps={{ text: 'Back' }}
+          useGridRef
+        />
+        {!skipLocationName && (
+          <MapHeader.LocationName
+            onChange={onLocationNameChange}
+            value={location.name}
+            icon={locationNameIcon}
+            placeholder="Site name eg nearby village"
+            suggestions={appModel.attrs.locations || []}
+            geocodingParams={{
+              access_token: config.map.mapboxApiKey,
+              types: 'locality,place,district,neighborhood,region,postcode',
+              country: 'GB',
+            }}
+          />
+        )}
+      </MapHeader>
 
-                setShowPastLocations(true);
-              }}
-              fill="clear"
-              color="dark"
-            >
-              <IonIcon src={starOutline} />
-            </IonButton>
-          </div>
-        </div>
+      <Main>
+        {isMapboxMap && (
+          <MapboxMap
+            location={location}
+            parentLocation={parentLocation}
+            isDisabled={isDisabled}
+            onMapClick={onMapClick}
+            onGPSClick={onGPSClick}
+            currentStyle={currentStyle}
+            onLayersClick={onLayersClick}
+            onPastLocationsClick={!skipPastLocations && onPastLocationsClick}
+            isLocating={model.gps.locating}
+          />
+        )}
 
-        <IonModal
-          id="bottom-sheet"
+        {!isMapboxMap && (
+          <LeafletMap
+            model={model}
+            location={location}
+            setLocation={setLocation}
+            onGPSClick={onGPSClick}
+            onLayersClick={onLayersClick}
+            onPastLocationsClick={!skipPastLocations && onPastLocationsClick}
+          />
+        )}
+
+        <MapSettingsPanel isOpen={showSettings} onClose={onCloseSettings}>
+          <RadioInput
+            options={styles}
+            onChange={onStyleChange}
+            value={currentStyle}
+            className="no-padding"
+          />
+        </MapSettingsPanel>
+
+        <PastLocationsPanel
           isOpen={showPastLocations}
-          backdropDismiss={false}
-          backdropBreakpoint={0.5}
-          breakpoints={SNAP_POSITIONS}
-          initialBreakpoint={DEFAULT_SNAP_POSITION}
-          canDismiss
-          onIonModalWillDismiss={() => {
-            setShowPastLocations(false);
-          }}
-        >
-          <IonHeader class="ion-no-border">
-            <IonToolbar />
-          </IonHeader>
-          <IonContent>
-            <PastLocationsList onSelect={onSelectPastLoaction} />
-          </IonContent>
-        </IonModal>
-      </ModelLocationOrig>
-    </>
+          model={model}
+          onClose={() => setShowPastLocations(false)}
+        />
+      </Main>
+    </Page>
   );
 };
 
-(ModelLocation as any).WithoutName = (props: Props) => (
-  <ModelLocation {...props} locationNameProps={{}} />
+(ModelLocationMap as any).WithoutName = (props: Props) => (
+  <ModelLocationMap {...props} skipLocationName />
 );
 
-export default ModelLocation;
+export default observer(ModelLocationMap);
