@@ -1,24 +1,34 @@
 /* eslint-disable import/prefer-default-export */
 export { default as ErrorBoundary } from '@flumens/ionic/dist/components/Main/ErrorBoundary';
 export { default as initAnalytics } from '@flumens/ionic/dist/utils/analytics';
-export { default as ModelLocation } from '@flumens/ionic/dist/components/ModelLocationMap';
 export { default as Main } from '@flumens/ionic/dist/components/Main';
 export { default as Page } from '@flumens/ionic/dist/components/Page';
 export { default as Header } from '@flumens/ionic/dist/components/Header';
 export { default as RouteWithModels } from '@flumens/ionic/dist/components/RouteWithModels';
 export {
   default as Attr,
-  Props as AttrProps,
+  type Props as AttrProps,
 } from '@flumens/ionic/dist/components/Attr';
 export {
   default as AttrPage,
-  Props as PageProps,
+  type Props as PageProps,
 } from '@flumens/ionic/dist/components/AttrPage';
+export {
+  default as MapContainer,
+  useMapStyles,
+} from '@flumens/ionic/dist/components/Map/Container';
+export { default as MapHeader } from '@flumens/ionic/dist/components/Map/Header';
+export { default as MapSettingsPanel } from '@flumens/ionic/dist/components/Map/SettingsPanel';
+export * from '@flumens/ionic/dist/components/Map/utils';
 export { default as Gallery } from '@flumens/ionic/dist/components/Gallery';
-export { default as RadioInput } from '@flumens/ionic/dist/components/RadioInput';
+export {
+  default as RadioInput,
+  type Option as RadioInputOption,
+} from '@flumens/ionic/dist/components/RadioInput';
 export { default as PhotoPicker } from '@flumens/ionic/dist/components/PhotoPicker';
 export { default as date } from '@flumens/ionic/dist/utils/date';
 export { default as device } from '@flumens/ionic/dist/utils/device';
+export * from '@flumens/ionic/dist/utils/uuid';
 export { useToast, useAlert, useLoader } from '@flumens/ionic/dist/hooks';
 export { default as Collapse } from '@flumens/ionic/dist/components/Collapse';
 export { default as InfoMessage } from '@flumens/ionic/dist/components/InfoMessage';
@@ -29,37 +39,39 @@ export { default as InfoButton } from '@flumens/ionic/dist/components/InfoButton
 export { default as Section } from '@flumens/ionic/dist/components/Section';
 export {
   default as MenuAttrItem,
-  Props as MenuAttrItemProps,
+  type Props as MenuAttrItemProps,
 } from '@flumens/ionic/dist/components/MenuAttrItem';
 export { default as MenuAttrItemFromModel } from '@flumens/ionic/dist/components/MenuAttrItemFromModel';
 export { default as MenuAttrToggle } from '@flumens/ionic/dist/components/MenuAttrToggle';
 export { default as ImageCropper } from '@flumens/ionic/dist/components/ImageCropper';
+export { default as ModelValidationMessage } from '@flumens/ionic/dist/components/ModelValidationMessage';
 export { default as Store } from '@flumens/ionic/dist/models/Store';
 export { default as initStoredSamples } from '@flumens/ionic/dist/models/initStoredSamples';
+export * from '@flumens/ionic/dist/models/Indicia/helpers';
 export {
   default as Model,
-  Attrs as ModelAttrs,
+  type Attrs as ModelAttrs,
 } from '@flumens/ionic/dist/models/Model';
 export {
   default as Sample,
-  Attrs as SampleAttrs,
-  Metadata as SampleMetadata,
-  Options as SampleOptions,
-  RemoteConfig,
+  type Attrs as SampleAttrs,
+  type Metadata as SampleMetadata,
+  type Options as SampleOptions,
+  type RemoteConfig,
 } from '@flumens/ionic/dist/models/Indicia/Sample';
 export {
   default as Media,
-  Attrs as MediaAttrs,
+  type Attrs as MediaAttrs,
 } from '@flumens/ionic/dist/models/Indicia/Media';
 export {
   default as Occurrence,
-  Attrs as OccurrenceAttrs,
-  Metadata as OccurrenceMetadata,
-  Options as OccurrenceOptions,
+  type Attrs as OccurrenceAttrs,
+  type Metadata as OccurrenceMetadata,
+  type Options as OccurrenceOptions,
 } from '@flumens/ionic/dist/models/Indicia/Occurrence';
 export {
   default as DrupalUserModel,
-  Attrs as DrupalUserModelAttrs,
+  type Attrs as DrupalUserModelAttrs,
 } from '@flumens/ionic/dist/models/DrupalUserModel';
 export { default as InputWithValidation } from '@flumens/ionic/dist/components/InputWithValidation';
 export { default as UserFeedbackRequest } from '@flumens/ionic/dist/components/UserFeedbackRequest';
@@ -70,6 +82,11 @@ export {
 } from '@flumens/ionic/dist/hooks/navigation';
 export * from '@flumens/ionic/dist/utils/errors';
 export * from '@flumens/ionic/dist/utils/location';
-export * from '@flumens/ionic/dist/utils/validation';
 export * from '@flumens/ionic/dist/utils/image';
 export * from '@flumens/ionic/dist/utils/type';
+export { default as ImageWithBackground } from '@flumens/ionic/dist/components/ImageWithBackground';
+
+export {
+  type default as ElasticOccurrence,
+  type Media as ElasticOccurrenceMedia,
+} from '@flumens/ionic/dist/models/Indicia/ElasticOccurrence.d';
