@@ -10,8 +10,6 @@ export type Attrs = ModelAttrs & {
   locations: any[];
   attrLocks: any;
   autosync: boolean;
-  useGridRef: boolean;
-  useGridMap: boolean;
   useTraining: boolean;
 
   useExperiments: boolean;
@@ -23,6 +21,7 @@ export type Attrs = ModelAttrs & {
   showSurveysDeleteTip: boolean;
   shownLongPressTip: boolean;
   shownLockingSwipeTip: boolean;
+  showPastLocationsTip: boolean;
   feedbackGiven: boolean;
   taxonGroupFilters: any[];
   searchNamesOnly: '' | 'scientific' | 'common';
@@ -42,8 +41,6 @@ export const defaults: Attrs = {
   locations: [],
   attrLocks: { default: {}, complex: {} },
   autosync: true,
-  useGridRef: true,
-  useGridMap: true,
   useTraining: false,
 
   useExperiments: false,
@@ -55,6 +52,7 @@ export const defaults: Attrs = {
   showSurveysDeleteTip: true,
   shownLongPressTip: false,
   shownLockingSwipeTip: false,
+  showPastLocationsTip: true,
   feedbackGiven: false,
   taxonGroupFilters: [],
   searchNamesOnly: '',
@@ -68,6 +66,8 @@ export const defaults: Attrs = {
 };
 
 export class AppModel extends Model {
+  // eslint-disable-next-line
+  // @ts-ignore
   attrs: Attrs = Model.extendAttrs(this.attrs, defaults);
 
   isAttrLocked: any; // from extension
@@ -82,7 +82,7 @@ export class AppModel extends Model {
 
   getAllLocks: any; // from extension
 
-  setLocation: any; // from extension
+  setLocation!: (newLocation: any) => void; // from extension
 
   removeLocation: any; // from extension
 
